requirements.txt
transformers>=4.35.0
datasets>=2.14.0
evaluate>=0.4.0
torch>=2.0.0
Pillow>=9.0.0
numpy>=1.24.0
jiwer>=3.0.0
groq>=0.5.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0